+++ alectrnn/common/network_generator.cpp
@@ -21,13 +21,13 @@ InEdge::InEdge() {
   weight = 0;
 }
 
-InEdge::InEdge(int source_node, double source_weight)
+InEdge::InEdge(int source_node, float source_weight)
   : source(source_node), weight(source_weight) {
 }
 
 // Creates a vector for all the in coming edges in an all-to-all network
 std::vector< std::vector<InEdge> > All2AllNetwork(std::size_t num_nodes,
-    const double *weights) {
+    const float *weights) {
   std::vector< std::vector<InEdge> > node_neighbors(num_nodes);
   for (std::size_t iii = 0; iii < num_nodes; iii++) {
     std::vector<InEdge> neighbors(num_nodes);
@@ -40,7 +40,7 @@ std::vector< std::vector<InEdge> > All2AllNetwork(std::size_t num_nodes,
 }
 
 std::vector< std::vector<InEdge> > All2AllNetwork(std::size_t num_nodes,
-    double default_weight) {
+    float default_weight) {
   std::vector< std::vector<InEdge> > node_neighbors(num_nodes);
   for (std::size_t iii = 0; iii < num_nodes; iii++) {
     std::vector<InEdge> neighbors(num_nodes);
@@ -53,7 +53,7 @@ std::vector< std::vector<InEdge> > All2AllNetwork(std::size_t num_nodes,
 }
 
 std::vector< std::vector<InEdge> > FullSensorNetwork(std::size_t num_sensors,
-    std::size_t num_nodes, const double *weights) {
+    std::size_t num_nodes, const float *weights) {
   std::vector<std::vector<InEdge> > node_sensors(num_nodes);
   for (std::size_t iii = 0; iii < num_nodes; iii++) {
     std::vector<InEdge> sensors(num_sensors);
@@ -66,7 +66,7 @@ std::vector< std::vector<InEdge> > FullSensorNetwork(std::size_t num_sensors,
 }
 
 std::vector< std::vector<InEdge> > FullSensorNetwork(std::size_t num_sensors,
-    std::size_t num_nodes, double default_weight) {
+    std::size_t num_nodes, float default_weight) {
   std::vector<std::vector<InEdge> > node_sensors(num_nodes);
   for (std::size_t iii = 0; iii < num_nodes; iii++) {
     std::vector<InEdge> sensors(num_sensors);
@@ -79,7 +79,7 @@ std::vector< std::vector<InEdge> > FullSensorNetwork(std::size_t num_sensors,
 }
 
 void FillFullSensorNetwork(std::vector< std::vector<InEdge> > &network,
-    const double *weights) {
+    const float *weights) {
   for (std::size_t iii = 0; iii < network.size(); iii++) {
     for (std::size_t jjj = 0; jjj < network[iii].size(); jjj++) {
       network[iii][jjj].weight = weights[iii * network[iii].size() + jjj];
@@ -88,7 +88,7 @@ void FillFullSensorNetwork(std::vector< std::vector<InEdge> > &network,
 }
 
 void FillAll2AllNetwork(std::vector< std::vector<InEdge> > &network,
-    const double *weights) {
+    const float *weights) {
   for (std::size_t iii = 0; iii < network.size(); iii++) {
     for (std::size_t jjj = 0; jjj < network[iii].size(); jjj++) {
       network[iii][jjj].weight = weights[iii * network[iii].size() + jjj];
